remove remaining polling code for channels, members & metadata
