+++ src/store/data/actions/index.ts
@@ -1,27 +1,23 @@
 import createChannel from "./createChannel";
 import createCommunity from "./createCommunity";
 import createExpression from "./createExpression";
-import getNeighbourhoodChannels from "./getNeighbourhoodChannels";
-import getNeighbourhoodMetadata from "./getNeighbourhoodMetadata";
-import getNeighbourhoodMembers from "./getNeighbourhoodMembers";
 import joinCommunity from "./joinCommunity";
 import updateCommunity from "./updateCommunity";
 import showMessageNotification from "./showMessageNotification";
 import fetchNeighbourhoodMembers from "./fetchNeighbourhoodMembers";
 import fetchNeighbourhoodMetadata from "./fetchNeighbourhoodMetadata";
+import fetchNeighbourhoodChannels from "./fetchNeighbourhoodChannels";
 import joinChannelNeighbourhood from "./joinChannelNeighbourhood";
 
 export default {
   createChannel,
   createCommunity,
   createExpression,
-  getNeighbourhoodChannels,
-  getNeighbourhoodMetadata,
-  getNeighbourhoodMembers,
   joinCommunity,
   updateCommunity,
   showMessageNotification,
   fetchNeighbourhoodMembers,
   fetchNeighbourhoodMetadata,
+  fetchNeighbourhoodChannels,
   joinChannelNeighbourhood,
 };
